net, dhcp: Fix test_dhcp_broadcast[#ovs-bridge#]
The test was failing because the VM received an IPv4 address from an
unknown source on the `172.31.0.243/24` network, might be different DHCP
server in the cluster or from the VM itself, instead of the test’s intended DHCP server.

Fix: Added a unique DHCP client identifier so that the test’s DHCP
server can correctly recognize the client and assign an IPv4 address
from the range defined within the test.

Signed-off-by: Asia Khromov <azhivovk@redhat.com>