net, sriov: Refactor wait_for_ready_sriov_nodes
Problem:
SriovNetworkNodeState sometimes does not change to InProgress after sriov configuration and stays
in Succeeded state. This could lead to errors in setup and teardown of each node and declaration
of the error might take 1000 seconds (~17 minutes).

Fix:
Waiting for SNNS status to be updated according to the test's SriovNetworkNodePolicy instead of
waiting for state to change to InProgress.

Signed-off-by: Asia Khromov <azhivovk@redhat.com>