net, stability: Refactor dual_stack_network_data
Problem:
Some tests from single-stack IPv6 lane fail for `no ssh session`.
On single stack IPv6 clusters, by default, the primary interface is not configured
with IPv6 address so it must be configured manually. The existing configuration is
conflicting as it assigns static address for `gateway6` while the default for `dhcp6`
is True - this is breaking the assignment of IPv6 address on the primary interface of guest
and ssh is failing.

Fix:
Fix the conflicting configuration for primary interface when using IPv6 cluster by
refactoring the fixture `dual_stack_network_data` since this configuration must be
used both on dual-stack clusters and single-stack IPv6 clusters.
The new factory fixture is to be used for secondary interfaces
configuration in following PRs.

Signed-off-by: Asia Khromov <azhivovk@redhat.com>