net, migration: Split test_migration_with_masquerade
`test_migration_with_masquerade` test is collected to
run twice: once with `ip_family="ipv4"` and again with
`ip_family="ipv6"`.
On IPv6 single-stack lane it fails when it runs with
`ip_family="ipv4"`: `failed on setup with
"Failed: IPv4 is not supported in this cluster"`

This test must be split with parametrize to ensure it runs
with both ip families on dual-stack clusters and with IPv6
only on IPv6 single-stack clusters.

Signed-off-by: Asia Khromov <azhivovk@redhat.com>